+++ .gitignore
@@ -4,3 +4,4 @@ node_modules/
 snow-monkey-blocks.zip
 snow-monkey-blocks
 dist/
+npm-debug.log
