+++ App/RestAPI/BlockTemplatesCategories/config.php
@@ -1,0 +1,31 @@
+<?php
+/**
+ * @package snow-monkey-blocks
+ * @author inc2734
+ * @license GPL-2.0+
+ */
+
+use Snow_Monkey\Plugin\Blocks\App\RestAPI\Helper;
+
+return [
+	[
+		'title' => _x( 'Headings', 'template-category', 'snow-monkey-blocks' ),
+		'slug'  => 'smb-template-headings',
+		'isPro' => false,
+	],
+	[
+		'title' => _x( 'Features', 'template-category', 'snow-monkey-blocks' ),
+		'slug'  => 'smb-template-features',
+		'isPro' => false,
+	],
+	[
+		'title' => _x( 'Pricing', 'template-category', 'snow-monkey-blocks' ),
+		'slug'  => 'smb-template-pricing',
+		'isPro' => false,
+	],
+	[
+		'title' => _x( 'FAQ', 'template-category', 'snow-monkey-blocks' ),
+		'slug'  => 'smb-template-faq',
+		'isPro' => false,
+	],
+];
