+++ block/categories-list/view.php
@@ -1,7 +1,8 @@
 <?php
 /**
  * @package snow-monkey-blocks
- * @author kmix-39
+ * @author inc2734
+ * @contributor kmix-39
  * @license GPL-2.0+
  */
